[DEBUG] Issue with coherent wave in Paaschens
renormalized the Paasschens validation routine so as to obtain energies integrated over all angles for easier comparison with our code. Also, the coherent wave was not captured appropriately because x=ct was enforced too strongly. Modified to choose the x closest to ct in order to ensure that the coherent wave always appear in the simulation, whatever the space-time discretization